Worked around a GDC bug, kind of.
Before this commit running "./build.d --dflags "-m32" on Linux
resulted in internal compiler error. Apparently this depends
on the order of files on gdmd command line.

+++ build.d
@@ -451,7 +451,7 @@ void doit(string[] args)
             Compiler.LDC : "ldc2"][dc];
    
     if(types == [])
-        types = ["float", "double", "real"];
+        types = ["double", "float", "real"];
 
     if(simdOpt == "")
         simdOpt = dc == Compiler.GDMD && isLinux ? "sse-avx" : "sse";
